requirements.txt
annotated-types==0.7.0
anyio==4.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
exceptiongroup==1.2.2
fastapi==0.115.12
firebase-admin==6.8.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
msgpack==1.1.0
packaging==25.0
pip-autoremove==0.10.0
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.34.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
python-decouple==3.8
requests==2.32.3
rsa==4.9.1
setuptools==80.7.1
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
